Phase 2 step 5: Extract persistence
